requirements.txt
djangorestframework
APScheduler==3.7.0
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4